Improve lexemme collection in lexer
This commit pushes the process of collecting a lexemme to its own
function allowing it to be reused more easily, promoting code reuse.

This function also checks for potential buffer overflows, meaning we now
report errors when trying to collect a lexemme which excedes the length
of our lexemme buffer in bytes.